# docs/swagger.yaml
basePath: /
definitions:
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
  handler.signInInput:
    properties:
      login:
        type: string
      password:
        type: string
    required:
    - login
    - password
    type: object
  models.Client:
    properties:
      id:
        type: integer
      login:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      registrationDate:
        type: string
      surname:
        type: string
    required:
    - login
    - name
    - password
    - phone
    - surname
    type: object
  models.Manager:
    properties:
      login:
        type: string
      name:
        type: string
      password:
        type: string
      surname:
        type: string
    required:
    - login
    - name
    - password
    - surname
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: API Server for Delivery company Application
  license: {}
  title: Delivery company
  version: "1.0"
paths:
  /auth/client-sign-up:
    post:
      consumes:
      - application/json
      description: creat client account
      operationId: creat-client-account
      parameters:
      - description: account info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Client'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Client SignUp
      tags:
      - Auth
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: login to your personal account
      operationId: login-account
      parameters:
      - description: account info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.signInInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: SignIn
      tags:
      - Auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: create manager account
      operationId: create-manager-account
      parameters:
      - description: account info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Manager'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Manager SignUp
      tags:
      - Auth
  /client-api/cabinet/:
    get:
      description: getting information about the client
      operationId: get-info-client
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get info client
      tags:
      - Client
    put:
      consumes:
      - application/json
      description: update information about the client
      operationId: update-info-client
      parameters:
      - description: new info client
        in: body
        name: client
        required: true
        schema:
          $ref: '#/definitions/models.Client'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: update info client
      tags:
      - Client
...
  /manager-api/cabinet/:
    get:
      description: getting information about the manager
      operationId: get-info-manager
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get info manager
      tags:
      - Manager
    put:
      consumes:
      - application/json
      description: update information about the manager
      operationId: update-info-manager
      parameters:
      - description: new info manager
        in: body
        name: manager
        required: true
        schema:
          $ref: '#/definitions/models.Manager'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: update info manager
      tags:
      - Manager
  /manager-api/client/:
    get:
      description: Get all clietn by a manager
      operationId: get-all-client-manager
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all client
      tags:
      - Manager
  /manager-api/driver/:
    get:
      description: Get all driver by a manager
      operationId: get-all-driver-manager
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get all driver
      tags:
      - Manager
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"